Xong quen va doi mat khau; da co profile, co dashboard; da fix tim kiem, trang thai va them nut view trong quan ly import order cua nhan vien; da fix tao don nhap hang

+++ src/main/java/vn/cosbeauty/repository/SupplierRepository.java
@@ -3,5 +3,5 @@ package vn.cosbeauty.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.cosbeauty.entity.Supplier;
 
-public interface SupplierRepository extends JpaRepository<Supplier, Long> {
+public interface SupplierRepository extends JpaRepository<Supplier, Integer> {
 }
